fix template_obj.py for currency conversion properties.  not in a nice way, however there is not much of a choice.
the j.data.types.numeric.bytes2cur function requires PACKED BYTES
to be passed in... NOT a decoded (pretty-formatted) string like "10 USD".

so, as a work-around, the value is converted **TO** bytes (using
j.data.types.numeric.str2bytes) which is of course really inefficient
and THEN it can be passed in to the bytes2cur function.

really what should be happening is that instead of a string, a class
instance should be returned that points to the packed value, that
happens to have an __str__ (and/or __repr__) method.  although... having
a class instance is not particularly memory-efficient either...

unit tests have also been added.